+++ packages/grafana-data/src/datetime/index.ts
@@ -1,0 +1,5 @@
+// Names are too general to export globally
+import * as dateMath from './datemath';
+import * as rangeUtil from './rangeutil';
+export * from './moment_wrapper';
+export { dateMath, rangeUtil };
