+++ packages/grafana-data/src/index.ts
@@ -3,3 +3,4 @@ export * from './types';
 export * from './vector';
 export * from './dataframe';
 export * from './transformations';
+export * from './datetime';
